+++ views/admin/modules/view.php
@@ -17,7 +17,12 @@ use yii\widgets\DetailView;
         'attributes' => [
             'id',
             'name:ntext',
-            'description:ntext',
+            [
+                'attribute' => 'description',
+                'value' => function($data) use ($module) {
+                    return Yii::t('app/modules/'.$data->module, $data->description);
+                }
+            ],
 
             'class:ntext',
             // 'bootstrap:ntext',
@@ -83,10 +88,53 @@ use yii\widgets\DetailView;
             'protected:ntext',
             'priority:ntext',
 
-            'created_at:datetime',
-            'created_by:ntext',
-            'updated_at:datetime',
-            'updated_by:ntext',
+            [
+                'attribute' => 'created_at',
+                'label' => Yii::t('app/modules/admin','Created'),
+                'format' => 'html',
+                'value' => function($data) {
+
+                    $output = "";
+                    if ($user = $data->createdBy) {
+                        $output = Html::a($user->username, ['users/view', 'id' => $user->id], [
+                            'target' => '_blank',
+                            'data-pjax' => 0
+                        ]);
+                    } else if ($data->created_by) {
+                        $output = $data->created_by;
+                    }
+
+                    if (!empty($output))
+                        $output .= ", ";
+
+                    $output .= Yii::$app->formatter->format($data->created_at, 'datetime');
+                    return $output;
+                }
+            ],
+
+            [
+                'attribute' => 'updated_at',
+                'label' => Yii::t('app/modules/admin','Updated'),
+                'format' => 'html',
+                'value' => function($data) {
+
+                    $output = "";
+                    if ($user = $data->updatedBy) {
+                        $output = Html::a($user->username, ['users/view', 'id' => $user->id], [
+                            'target' => '_blank',
+                            'data-pjax' => 0
+                        ]);
+                    } else if ($data->updated_by) {
+                        $output = $data->updated_by;
+                    }
+
+                    if (!empty($output))
+                        $output .= ", ";
+
+                    $output .= Yii::$app->formatter->format($data->updated_at, 'datetime');
+                    return $output;
+                }
+            ],
 
         ],
     ]) ?>